+++ src/main/kotlin/ar/edu/unsam/model/TableData.kt
@@ -1,24 +1,22 @@
 package ar.edu.unsam.model
 abstract class BaseDataDTO(
-  open val productoId: Long,
+  open val productoId: Int,
   open val productoNombre: String,
   open val gondolaNombre: String,
   open val presentacionDesc: String
 )
 
 data class RepositorDataDTO(
-  val repositorId: Long,
-  // Campos heredados de ar.edu.unsam.ar.edu.unsam.model.BaseDataDTO
-  override val productoId: Long,
+  val repositorId: Int,
+  override val productoId: Int,
   override val productoNombre: String,
   override val gondolaNombre: String,
   override val presentacionDesc: String
 ) : BaseDataDTO(productoId, productoNombre, gondolaNombre, presentacionDesc)
 
 data class SectorDataDTO(
-  val sectorId: Long,
-  // Campos heredados de ar.edu.unsam.ar.edu.unsam.model.BaseDataDTO
-  override val productoId: Long,
+  val sectorId: Int,
+  override val productoId: Int,
   override val productoNombre: String,
   override val gondolaNombre: String,
   override val presentacionDesc: String
